Create HeatMappa0.1
Small basic script that allows to generate gene expression heatmaps. The script has been created with the tkinter package that allows to generate a small use friendly prompt that allows to load the file and save the generated heatmap as image in a user chosen location. This is a basic version that will be implemented in the future