Update HeatMappa_v1.0.0
This is a small python script that allows the generation of gene expression cluster heatmap. The tkinter package generates small prompt that allows the selection of the file, set some heatmap parameters and chose the heatmap saving location. This version do not give that much freedom but it will be implemented in the future.